Create Code
Spatial enhancement using transformation and image filtering is one the most widely
preprocessing tool in image processing based applications. Image filtering further is being used
to compliment segmentation and feature extraction process.
In this assignment, the aim is to give you a hands on experience of using these operations for a
real-life problem. Fundus images are digital images of human retinal which are used to diagnose
different retinal diseases. The main landmarks on fundus images are optic disc, vessels and fovea
(macula) as shown in following image
Your aim in this assignment is to use your knowledge about spatial enhancement and connected
component analysis to extract optic disc from given images. You are given a set of 50 fundus
images containing original colored fundus images along with their respective blood vessel map (as
shown in figure below) and original optic disc location in form of (row, col).
You need to extract optic disc from all these 50 images and find out error in each image and
overall average error. The images for this assignment can be accessed using following link
https://drive.google.com/drive/folders/1DxmL9I2772qTCYwlbMk1KpKPtJb85o-H?usp=sharing
Deliverables are
1. Algorithm for extraction of bright regions from input fundus images: As you can see that
optic disc is a bright region in comparison to its background so first step should be looking for 
such bright regions.
2. Algorithm to identify optic disc location: Now in case of normal images there should be only 
1 bright region that is optic disc. But in abnormal case there are other bright lesions
(exudates, as shown in first figure) so it would extract optic disc along with other false regions 
due to these bright lesions. Now you have to device a method to find out true optic disc and 
get coordinates for that. Blood vessels map can help here as vessels are originated from optic 
disc. Also, mark the location as shown in figure below
3. Error Calculation: Now as you have extracted optic disc along with its coordinates (row
and col), so we can find out error between actual Optic disc location and location which you
have extracted. For this, use Euclidean distance that is
Here true location of optic disc is represented as (x1, y1) and the one which you have extracted 
is represented as (x2, y2). Calculate this error for each image and add in your report in a tabular 
manner.